Move buttons into form elements component
